requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.25.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
huggingface-hub==0.19.4 